+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## TBD
 
+* Generate shared object mapping files for libunity and libil2cpp
+  [#312](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/312)
+
 * Register unity shared object generation and upload tasks
   [#311](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/311)